Adapt to use the new UITestCase class
The test should be otherwise identical.
